+++ kubernetes-client-api/src/main/java/io/fabric8/kubernetes/client/utils/KubernetesResourceUtil.java
@@ -18,6 +18,8 @@ package io.fabric8.kubernetes.client.utils;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import io.fabric8.kubernetes.api.builder.VisitableBuilder;
+import io.fabric8.kubernetes.api.model.ConfigMap;
+import io.fabric8.kubernetes.api.model.ConfigMapBuilder;
 import io.fabric8.kubernetes.api.model.DefaultKubernetesResourceList;
 import io.fabric8.kubernetes.api.model.EnvVar;
 import io.fabric8.kubernetes.api.model.EnvVarBuilder;
@@ -31,9 +33,18 @@ import io.fabric8.kubernetes.api.model.Secret;
 import io.fabric8.kubernetes.api.model.SecretBuilder;
 import io.fabric8.kubernetes.client.readiness.Readiness;
 
+import java.io.File;
+import java.io.IOException;
+import java.nio.ByteBuffer;
+import java.nio.charset.CharacterCodingException;
+import java.nio.charset.CodingErrorAction;
 import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.nio.file.LinkOption;
+import java.nio.file.Path;
 import java.time.Duration;
 import java.time.Instant;
+import java.util.AbstractMap;
 import java.util.ArrayList;
 import java.util.Base64;
 import java.util.Collections;
@@ -42,7 +53,10 @@ import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.regex.Pattern;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public class KubernetesResourceUtil {
   private KubernetesResourceUtil() {
@@ -455,6 +469,142 @@ public class KubernetesResourceUtil {
     String dockerConfigAsStr = Serialization.jsonMapper().writeValueAsString(dockerConfigMap);
 
     return createDockerSecret(secretName, dockerConfigAsStr);
+  }
+
+  /**
+   * Create new ConfigMap from files/directories
+   *
+   * @param name name of Configmap to create
+   * @param dirOrFilePaths a var-arg for directory of file paths.
+   * @return ConfigMap with data as key-value pair of file names and their contents
+   * @throws IOException in case of failure while reading file
+   */
+  public static ConfigMap createConfigMapFromDirOrFiles(final String name, final Path... dirOrFilePaths)
+      throws IOException {
+    ConfigMapBuilder configMapBuilder = new ConfigMapBuilder();
+    configMapBuilder.withNewMetadata().withName(name).endMetadata();
+    for (Path dirOrFilePath : dirOrFilePaths) {
+      final File file = dirOrFilePath.toFile();
+      if (!file.exists()) {
+        throw new IllegalArgumentException("invalid file path provided " + dirOrFilePath);
+      }
+      addEntriesFromDirOrFileToConfigMap(configMapBuilder, file.getName(), dirOrFilePath);
+    }
+    return configMapBuilder.build();
+  }
+
+  /**
+   * Create new ConfigMap with specified entry in its data field
+   *
+   * @param key key in ConfigMap's data map
+   * @param value path to a file or a directory whose contents would be read as value
+   * @return a ConfigMap whose data contains provided key and value
+   * @throws IOException in case of failure while reading file or directory
+   */
+  public static ConfigMap createNewConfigMapWithEntry(final String key, final Path value) throws IOException {
+    ConfigMapBuilder configMapBuilder = new ConfigMapBuilder();
+    addEntriesFromDirOrFileToConfigMap(configMapBuilder, key, value);
+    return configMapBuilder.build();
+  }
+
+  /**
+   * Merge ConfigMap data of two ConfigMaps
+   *
+   * @param cm1 first ConfigMap
+   * @param cm2 cm2 Configmap which would be modified
+   * @return ConfigMap containing data of both ConfigMaps
+   */
+  public static ConfigMap mergeConfigMapData(final ConfigMap cm1, final ConfigMap cm2) {
+    ConfigMapBuilder resultConfigMapBuilder = new ConfigMapBuilder();
+    if (cm1 != null || cm2 != null) {
+      if (cm1 == null) {
+        return cm2;
+      } else if (cm2 == null) {
+        return cm1;
+      } else {
+        Map<String, String> mergedData = mergeMaps(cm1.getData(), cm2.getData());
+        Map<String, String> mergedBinaryData = mergeMaps(cm1.getBinaryData(), cm2.getBinaryData());
+        resultConfigMapBuilder.withData(mergedData);
+        resultConfigMapBuilder.withBinaryData(mergedBinaryData);
+      }
+    }
+    return resultConfigMapBuilder.build();
+  }
+
+  private static Map<String, String> mergeMaps(Map<String, String> m1, Map<String, String> m2) {
+    if (m1 == null && m2 == null) {
+      return Collections.emptyMap();
+    } else if (m1 == null) {
+      return m2;
+    } else if (m2 == null) {
+      return m1;
+    } else {
+      return Stream.of(m1, m2)
+          .flatMap(m -> m.entrySet().stream())
+          .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
+    }
+  }
+
+  private static Map.Entry<String, String> createConfigMapEntry(final String key, final Path file) throws IOException {
+    final byte[] bytes = Files.readAllBytes(file);
+    if (isFileWithBinaryContent(file)) {
+      final String value = Base64.getEncoder().encodeToString(bytes);
+      return new AbstractMap.SimpleEntry<>(key, value);
+    } else {
+      return new AbstractMap.SimpleEntry<>(key, new String(bytes));
+    }
+  }
+
+  private static boolean isFileWithBinaryContent(final Path file) throws IOException {
+    final byte[] bytes = Files.readAllBytes(file);
+    try {
+      StandardCharsets.UTF_8.newDecoder()
+          .onMalformedInput(CodingErrorAction.REPORT)
+          .onUnmappableCharacter(CodingErrorAction.REPORT)
+          .decode(ByteBuffer.wrap(bytes));
+      return false;
+    } catch (CharacterCodingException e) {
+      return true;
+    }
+  }
+
+  private static void addEntriesFromDirectoryToConfigMap(ConfigMapBuilder configMapBuilder, final Path path)
+      throws IOException {
+    try (Stream<Path> files = Files.list(path)) {
+      files.filter(p -> !Files.isDirectory(p, LinkOption.NOFOLLOW_LINKS)).forEach(file -> {
+        try {
+          addEntryToConfigMap(configMapBuilder, createConfigMapEntry(file.getFileName().toString(), file), file);
+        } catch (IOException e) {
+          throw new IllegalArgumentException(e);
+        }
+      });
+    }
+  }
+
+  private static void addEntryFromFileToConfigMap(ConfigMapBuilder configMapBuilder, final String key,
+      final Path file) throws IOException {
+    String entryKey = Optional.ofNullable(key).orElse(file.toFile().getName());
+    Map.Entry<String, String> configMapEntry = createConfigMapEntry(entryKey, file);
+    addEntryToConfigMap(configMapBuilder, configMapEntry, file);
+  }
+
+  private static void addEntryToConfigMap(ConfigMapBuilder configMapBuilder, Map.Entry<String, String> entry,
+      final Path file)
+      throws IOException {
+    if (isFileWithBinaryContent(file)) {
+      configMapBuilder.addToBinaryData(entry.getKey(), entry.getValue());
+    } else {
+      configMapBuilder.addToData(entry.getKey(), entry.getValue());
+    }
+  }
+
+  private static void addEntriesFromDirOrFileToConfigMap(ConfigMapBuilder configMapBuilder, final String key,
+      final Path dirOrFilePath) throws IOException {
+    if (Files.isDirectory(dirOrFilePath, LinkOption.NOFOLLOW_LINKS)) {
+      addEntriesFromDirectoryToConfigMap(configMapBuilder, dirOrFilePath);
+    } else {
+      addEntryFromFileToConfigMap(configMapBuilder, key, dirOrFilePath);
+    }
   }
 
   private static Map<String, Object> createDockerRegistryConfigMap(String dockerServer, String username, String password) {