+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/impl/KubernetesClientImpl.java
@@ -101,6 +101,7 @@ import io.fabric8.kubernetes.client.dsl.DiscoveryAPIGroupDSL;
 import io.fabric8.kubernetes.client.dsl.EventingAPIGroupDSL;
 import io.fabric8.kubernetes.client.dsl.ExtensionsAPIGroupDSL;
 import io.fabric8.kubernetes.client.dsl.FlowControlAPIGroupDSL;
+import io.fabric8.kubernetes.client.dsl.FromFileCreatableOperation;
 import io.fabric8.kubernetes.client.dsl.FunctionCallable;
 import io.fabric8.kubernetes.client.dsl.InOutCreateable;
 import io.fabric8.kubernetes.client.dsl.MetricAPIGroupDSL;
@@ -141,6 +142,7 @@ import io.fabric8.kubernetes.client.dsl.internal.apps.v1.ReplicaSetOperationsImp
 import io.fabric8.kubernetes.client.dsl.internal.apps.v1.StatefulSetOperationsImpl;
 import io.fabric8.kubernetes.client.dsl.internal.batch.v1.JobOperationsImpl;
 import io.fabric8.kubernetes.client.dsl.internal.certificates.v1.CertificateSigningRequestOperationsImpl;
+import io.fabric8.kubernetes.client.dsl.internal.core.v1.ConfigMapOperationsImpl;
 import io.fabric8.kubernetes.client.dsl.internal.core.v1.PodOperationsImpl;
 import io.fabric8.kubernetes.client.dsl.internal.core.v1.ReplicationControllerOperationsImpl;
 import io.fabric8.kubernetes.client.dsl.internal.core.v1.ServiceOperationsImpl;
@@ -239,6 +241,7 @@ public class KubernetesClientImpl extends BaseClient implements NamespacedKubern
     this.getHandlers().register(Pod.class, PodOperationsImpl::new);
     this.getHandlers().register(Job.class, JobOperationsImpl::new);
     this.getHandlers().register(Service.class, ServiceOperationsImpl::new);
+    this.getHandlers().register(ConfigMap.class, ConfigMapOperationsImpl::new);
     this.getHandlers().register(Deployment.class, DeploymentOperationsImpl::new);
     this.getHandlers().register(io.fabric8.kubernetes.api.model.extensions.Deployment.class,
         io.fabric8.kubernetes.client.dsl.internal.extensions.v1beta1.DeploymentOperationsImpl::new);
@@ -484,8 +487,8 @@ public class KubernetesClientImpl extends BaseClient implements NamespacedKubern
    * {@inheritDoc}
    */
   @Override
-  public MixedOperation<ConfigMap, ConfigMapList, Resource<ConfigMap>> configMaps() {
-    return resources(ConfigMap.class, ConfigMapList.class);
+  public FromFileCreatableOperation<ConfigMap, ConfigMapList, Resource<ConfigMap>> configMaps() {
+    return new ConfigMapOperationsImpl(this);
   }
 
   /**